Mouvement masi problèmes avec les bords de l'écran lorsque size>1

+++ include/CellularUnit.hpp
@@ -10,21 +10,22 @@ class CellularUnit
 {
 
 public:
-    CellularUnit(int x, int y);
+    CellularUnit(float x, float y);
     ~CellularUnit();
 
     void move();
 
-    inline std::pair<int, int> getCoords() const { return coords_; }
-    inline int getX() const { return coords_.first; }
-    inline int getY() const { return coords_.second; }
-    inline void setCoords(int x, int y) { coords_ = std::pair<int, int>(x, y); }
+    inline std::pair<float, float> getCoords() const { return coords_; }
+    inline float getX() const { return coords_.first; }
+    inline float getY() const { return coords_.second; }
+    inline void setCoords(float x, float y) { coords_ = std::pair<float, float>(x, y); }
     inline void setVelocity(float x, float y) { velocity_ = std::pair<float, float>(x, y); }
     inline std::pair<float, float> getVelocity() const { return velocity_; }
 
 
 private:
-    std::pair<int, int> coords_;
+    std::pair<float, float> coords_;
+    // Vecocity is a pair of float, the first float is the speed, the second is the angle
     std::pair<float, float> velocity_;
 
 };
